Refactoring

+++ distributor/devfield.cpp
@@ -2,25 +2,23 @@
 
 #include <iostream>
 
-//#define FMASK
-
 DevField::DevField(Sizes size, double d)
     : m_size(size)
 {
     m_div = uint8_t(ceil(7 / d));
     m_field = new uint8_t**[size.x * m_div];
 #ifdef FMASK
-    m_mask = new int8_t**[size.x * div];
+    m_mask = new int16_t**[size.x * div];
 #endif
     for (uint32_t ix = 0; ix < size.x * m_div; ++ix) {
         m_field[ix] = new uint8_t*[size.y * m_div];
 #ifdef FMASK
-        m_mask[ix] = new int8_t*[size.y * div];
+        m_mask[ix] = new int16_t*[size.y * div];
 #endif
         for (uint32_t iy = 0; iy < size.y * m_div; ++iy) {
             m_field[ix][iy] = new uint8_t[size.z * m_div];
 #ifdef FMASK
-            m_mask[ix][iy] = new int8_t[size.z * div];
+            m_mask[ix][iy] = new int16_t[size.z * div];
 #endif
             for (uint32_t iz = 0; iz < size.z * m_div; ++iz) {
                 m_field[ix][iy][iz] = d_empty;
